Move ray_utils.py from entrypoints to utils directory
