+++ python/sglang/srt/entrypoints/http_server_actor.py
@@ -148,7 +148,7 @@ class HttpServerActor:
     def __init__(self, **server_kwargs):
         from ray.util.scheduling_strategies import PlacementGroupSchedulingStrategy
 
-        from sglang.srt.entrypoints.ray_utils import create_placement_groups
+        from sglang.srt.utils.ray_utils import create_placement_groups
 
         pgs = create_placement_groups(
             tp_size=server_kwargs.get("tp_size", 1),
